Moved remaining sitenow:multisite:* functionality to uiowa:multisite:* and removed Sitenow Multisite command file.
